Fixed Parser
Parser now works completely. Need to add more functionality for incomplete messages/data. Also, the amount of component values that can be sent is a max of 4 atm since the array size must be defined. However, this can be changed if we decide we need a lot of values for a component. It might take longer to process.