Bar fade utility
If a bar has a fade bar (previously called barMid), the a "fade bar" will be displayed when the bar is decreased. The fade bar falloff speed is proportional to the percent lost in the primary bar, and will engage after a short delay. This will hopefully add gravitas to loss on bars such as the health bar.